// frontend/src/shared/api/generated/applications/models/invitationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Applications
 * OpenAPI spec version: 0.1.0
 */
import type { InvitationStatus } from './invitationStatus';
import type { Vacancy } from './vacancy';

export interface InvitationResponse {
  id: string;
  status: InvitationStatus;
  candidate_id: string;
  employer_id: string;
  vacancy: Vacancy;
  message: string | null;
  response_message: string | null;
  created_at: string;
  responded_at: string | null;
  expires_at: string;
}