// frontend/src/shared/api/generated/applications/models/invitationStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Applications
 * OpenAPI spec version: 0.1.0
 */

export type InvitationStatus = typeof InvitationStatus[keyof typeof InvitationStatus];


export const InvitationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
  withdrawn: 'withdrawn',
  expired: 'expired',
} as const;